move: services/apiのデータを一階層上に持ってきた。それに伴うimport時のpathの修正
