+++ src/__tests__/services/prefecture/fetchPrefectures.test.ts
@@ -1,4 +1,4 @@
-import fetchPrefectures from '@/services/api/prefecture/fetchPrefectures';
+import fetchPrefectures from '@/services/prefecture/fetchPrefectures';
 import { Prefecture } from '@/types/api/models/prefecture/Prefecture';
 
 /**
@@ -200,7 +200,7 @@ describe('fetchPrefectures', () => {
       },
     }));
 
-    const fetchPrefectures = (await import('@/services/api/prefecture/fetchPrefectures')).default;
+    const fetchPrefectures = (await import('@/services/prefecture/fetchPrefectures')).default;
 
     await expect(fetchPrefectures()).rejects.toThrow('API_ENDPOINT is empty');
   });
@@ -223,7 +223,7 @@ describe('fetchPrefectures', () => {
       },
     }));
 
-    const fetchPrefectures = (await import('@/services/api/prefecture/fetchPrefectures')).default;
+    const fetchPrefectures = (await import('@/services/prefecture/fetchPrefectures')).default;
 
     await expect(fetchPrefectures()).rejects.toThrow('X_API_KEY is empty');
   });
@@ -253,7 +253,7 @@ describe('fetchPrefectures', () => {
       },
     }));
 
-    const fetchPrefectures = (await import('@/services/api/prefecture/fetchPrefectures')).default;
+    const fetchPrefectures = (await import('@/services/prefecture/fetchPrefectures')).default;
 
     await expect(fetchPrefectures()).rejects.toThrow('PREFECTURES is empty');
   });
